app/requirements.txt
Flask==3.0.2
requests==2.31.0
gspread==6.1.4
google-api-core==2.18.0
google-api-python-client==2.160.0
google-cloud-secret-manager==2.22.1